+++ apps/mobile-app/auth_routes/farmacias.ts
