+++ apps/mobile-app/schemas/device.ts
