+++ apps/mobile-app/schemas/farmacias.ts
