+++ auth/routes/routes.ts
@@ -6,7 +6,7 @@ import { authUsers } from '../schemas/permisos';
 import * as authOrganizaciones from './../../core/tm/schemas/organizacion';
 import { profesional } from './../../core/tm/schemas/profesional';
 import * as mongoose from 'mongoose';
-import * as authMobile from '../../modules/mobileApp/controller/AuthController';
+import * as authMobile from '../../apps/mobile-app/controller/AuthController';
 
 const isReachable = require('is-reachable');
 let sha1Hash = require('sha1');
