+++ modules/rup/routes/prestacion.ts
@@ -8,7 +8,7 @@ import { model as PrestacionAdjunto } from '../schemas/prestacion-adjuntos';
 
 import { buscarPaciente } from '../../../core/mpi/controller/paciente';
 import * as frecuentescrl from '../controllers/frecuentesProfesional';
-import { NotificationService } from '../../mobileApp/controller/NotificationService';
+import { NotificationService } from '../../../apps/mobile-app/controller/NotificationService';
 
 import { iterate, convertToObjectId, buscarEnHuds, matchConcepts } from '../controllers/rup';
 import { Logger } from '../../../utils/logService';
